+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release history for s390-tools (MIT version)
   Changes of existing tools:
 
   Bug Fixes:
+  - mon_procd: fix parsing of /proc/<pid>/stat
 
  * __v2.5.0 (2018-06-08)__